- validated that products with past orders can't be deleted

+++ MoneyFellows.ProductOrder.Application/Products/Commands/DeleteProductCommand.cs
@@ -12,10 +12,11 @@ public class DeleteProductCommand : IRequest
 public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand>
 {
     private readonly IProductRepository _productRepository;
-
-    public DeleteProductCommandHandler(IProductRepository productRepository)
+    private readonly IOrderRepository _orderRepository;
+    public DeleteProductCommandHandler(IProductRepository productRepository, IOrderRepository orderRepository)
     {
         _productRepository = productRepository;
+        _orderRepository = orderRepository;
     }
     public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
@@ -23,6 +24,11 @@ public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand>
         if (product == null)
         {
             throw new Exception("Product with Id: " + request.Id + " doesn't exist");
+        }
+        var isProductInPastOrders = await _orderRepository.AnyAsync(o => o.DeliveryTime < DateTime.Now && o.OrderDetails.Any(od => od.ProductId == request.Id), request.Id);
+        if (isProductInPastOrders)
+        {
+            throw new Exception("Product with Id: " + request.Id + " cannot be deleted because it is associated with past orders.");
         }
 
         await _productRepository.DeleteAsync(request.Id);
